Fix Yoga Play, Add Mental Health, Leaderboard, Reminder

+++ app/src/main/java/com/bangkit23/hidupsehat/data/source/remote/response/UserNeedsResponse.kt
@@ -35,7 +35,7 @@ data class UserNeedsResponse(
 	val actualWater: Int? = null,
 
 	@field:SerializedName("waterNeeds")
-	val waterNeeds: Int? = null,
+	val waterNeeds: Double? = null,
 
 	@field:SerializedName("name")
 	val name: String? = null,
