+++ app/src/main/java/com/bangkit23/hidupsehat/domain/model/user/UserNeeds.kt
@@ -11,7 +11,7 @@ data class UserNeeds(
     val actualSleep: Int? = null,
     val imgUrl: String? = null,
     val actualWater: Int? = null,
-    val waterNeeds: Int? = null,
+    val waterNeeds: Double? = null,
     val name: String? = null,
     val contactNumber: String? = null,
     val calorieNeeds: Int? = null,
